[catalogue] Use int64 for book's ID

+++ services/catalogue/internal/domain/repositories/book_repository.go
@@ -8,5 +8,5 @@ import (
 
 type BookRepository interface {
 	Create(ctx context.Context, book entities.Book) error
-	GetByID(ctx context.Context, id int) (*entities.Book, error)
+	GetByID(ctx context.Context, id int64) (*entities.Book, error)
 }
